Move API definitions to -api, move server deps to -server
