+++ package.json
@@ -104,28 +104,13 @@
     "@babel/preset-env": "^7.15.6",
     "@babel/preset-react": "^7.8.3",
     "@babel/preset-typescript": "^7.15.0",
-    "@feathersjs/adapter-commons": "^4.5.11",
-    "@feathersjs/authentication": "^4.5.11",
-    "@feathersjs/authentication-local": "^4.5.11",
-    "@feathersjs/authentication-oauth": "^4.5.11",
-    "@feathersjs/errors": "^4.5.11",
-    "@feathersjs/express": "^4.5.11",
-    "@feathersjs/feathers": "^4.5.11",
     "@pmmmwh/react-refresh-webpack-plugin": "^0.5.1",
     "@sentry/browser": "^6.12.0",
     "@sentry/node": "^6.12.0",
     "@svgr/webpack": "5.5.0",
     "@types/babel-core": "^6.25.7",
-    "@types/compression": "^1.7.2",
-    "@types/debug": "^4.1.7",
-    "@types/email-templates": "^8.0.4",
-    "@types/express-sslify": "^1.2.2",
     "@types/lodash": "^4.14.173",
-    "@types/mailgun-js": "^0.22.12",
-    "@types/node": "^16.9.2",
-    "@types/node-cron": "^2.0.4",
     "@types/node-sass": "^4.11.2",
-    "@types/nodemailer-mailgun-transport": "^1.4.3",
     "@types/react-csv": "^1.1.2",
     "@types/react-dom": "^16.9.14",
     "@types/react-redux": "^7.1.18",
@@ -140,20 +125,15 @@
     "babel-loader": "8.1.0",
     "babel-plugin-named-asset-import": "^0.3.7",
     "babel-preset-react-app": "^10.0.0",
-    "base32-encode": "^1.2.0",
     "better-npm-run": "^0.1.1",
     "bfj": "^7.0.2",
     "bootstrap": "^4.6.0",
     "camelcase": "^6.1.0",
     "case-sensitive-paths-webpack-plugin": "2.3.0",
-    "compression": "^1.7.4",
-    "connect-history-api-fallback": "^1.6.0",
     "connected-react-router": "^6.7.0",
     "css-loader": "4.3.0",
-    "debug": "^4.3.2",
     "dotenv": "^10.0.0",
     "dotenv-expand": "5.1.0",
-    "email-templates": "^8.0.8",
     "eslint": "^7.32.0",
     "eslint-config-airbnb": "^18.2.1",
     "eslint-config-airbnb-typescript": "^14.0.0",
@@ -165,10 +145,6 @@
     "eslint-plugin-react-hooks": "^4.2.0",
     "eslint-plugin-simple-import-sort": "^7.0.0",
     "eslint-webpack-plugin": "^2.5.2",
-    "express": "^4.17.1",
-    "express-sslify": "^1.2.0",
-    "feathers-hooks-common": "^5.0.5",
-    "feathers-sequelize": "^6.2.0",
     "file-loader": "6.1.1",
     "formik": "^2.2.9",
     "fs-extra": "^9.0.1",
@@ -181,18 +157,9 @@
     "jest-watch-typeahead": "^0.6.4",
     "lerna": "^4.0.0",
     "lodash": "^4.17.15",
-    "mailgun-js": "^0.22.0",
     "mini-css-extract-plugin": "0.11.3",
-    "moment": "^2.24.0",
-    "moment-timezone": "^0.5.28",
-    "mysql2": "^2.3.0",
-    "node-cron": "^3.0.0",
     "node-sass": "^6.0.1",
-    "nodemailer": "^6.6.3",
-    "nodemailer-mailgun-transport": "^2.1.3",
     "optimize-css-assets-webpack-plugin": "5.0.4",
-    "pg": "^8.7.1",
-    "pg-hstore": "^2.3.3",
     "pnp-webpack-plugin": "1.6.4",
     "postcss-flexbugs-fixes": "4.2.1",
     "postcss-loader": "3.0.0",
@@ -225,7 +192,6 @@
     "rimraf": "^3.0.2",
     "sass-loader": "^10.0.5",
     "semver": "^7.3.5",
-    "sequelize": "^6.6.5",
     "stackframe": "^1.1.1",
     "style-loader": "1.3.0",
     "terser-webpack-plugin": "4.2.3",
