+++ packages/ilmomasiina-api/src/models/answer.ts
@@ -1,0 +1,9 @@
+import QuestionAttributes from './question';
+import SignupAttributes from './signup';
+
+export default interface AnswerAttributes {
+  id: string;
+  answer: string;
+  questionId: QuestionAttributes['id'];
+  signupId: SignupAttributes['id'];
+}
