+++ packages/ilmomasiina-api/src/models/user.ts
@@ -1,0 +1,5 @@
+export default interface UserAttributes {
+  id: number;
+  email: string;
+  password: string;
+}
