+++ packages/ilmomasiina-api/src/services/admin/slug/get.ts
@@ -1,0 +1,6 @@
+import EventAttributes from '../../../models/event';
+
+export interface AdminCheckSlugResponse {
+  id: EventAttributes['id'] | null;
+  title: string | null;
+}
