+++ packages/ilmomasiina-api/src/services/admin/slug/index.ts
@@ -1,0 +1,8 @@
+import { StringifyApi } from '../../../utils';
+import { AdminCheckSlugResponse } from './get';
+
+export type AdminSlugServiceResponses = AdminCheckSlugResponse;
+
+export namespace AdminSlug {
+  export type Check = StringifyApi<AdminCheckSlugResponse>;
+}
