+++ packages/ilmomasiina-api/tsconfig.json
@@ -18,18 +18,13 @@
     "resolveJsonModule": true,
     "isolatedModules": true,
     "incremental": true,
-    "outDir": "dist",
-    "rootDirs": [
-      "src",
-      "config"
-    ],
-    "noEmit": true
+    "outDir": "dist"
   },
   "exclude": [
     "node_modules",
     "dist"
   ],
   "include": [
-    "src"
+    "src/**/*"
   ]
 }
