+++ packages/ilmomasiina-frontend/package.json
@@ -1,8 +1,7 @@
 {
   "name": "@tietokilta/ilmomasiina-frontend",
   "version": "2.0.0-alpha1",
-  "scripts": {
-  },
+  "scripts": {},
   "repository": {
     "type": "git",
     "url": "git+https://github.com/Tietokilta/ilmomasiina.git"
@@ -13,6 +12,7 @@
   },
   "homepage": "https://github.com/Tietokilta/ilmomasiina#readme",
   "dependencies": {
-    "@tietokilta/ilmomasiina-api": "2.0.0-alpha1"
+    "@tietokilta/ilmomasiina-api": "2.0.0-alpha1",
+    "better-npm-run": "^0.1.1"
   }
 }
