+++ packages/ilmomasiina-server/package.json
@@ -1,18 +1,60 @@
 {
   "name": "@tietokilta/ilmomasiina-server",
   "version": "2.0.0-alpha1",
-  "scripts": {
-  },
+  "license": "MIT",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/Tietokilta/ilmomasiina.git"
   },
-  "license": "MIT",
   "bugs": {
     "url": "https://github.com/Tietokilta/ilmomasiina/issues"
   },
   "homepage": "https://github.com/Tietokilta/ilmomasiina#readme",
+  "scripts": {
+    "build": "tsc",
+    "clean": "rimraf dist"
+  },
   "dependencies": {
-    "@tietokilta/ilmomasiina-api": "2.0.0-alpha1"
+    "@feathersjs/adapter-commons": "^4.5.11",
+    "@feathersjs/authentication": "^4.5.11",
+    "@feathersjs/authentication-local": "^4.5.11",
+    "@feathersjs/errors": "^4.5.11",
+    "@feathersjs/express": "^4.5.11",
+    "@feathersjs/feathers": "^4.5.11",
+    "@tietokilta/ilmomasiina-api": "2.0.0-alpha1",
+    "@types/compression": "^1.7.2",
+    "@types/debug": "^4.1.7",
+    "@types/email-templates": "^8.0.4",
+    "@types/express-sslify": "^1.2.2",
+    "@types/lodash": "^4.14.175",
+    "@types/node": "^16.9.2",
+    "@types/node-cron": "^2.0.4",
+    "@types/nodemailer": "^6.4.4",
+    "@types/nodemailer-mailgun-transport": "^1.4.3",
+    "base32-encode": "^2.0.0",
+    "better-npm-run": "^0.1.1",
+    "compression": "^1.7.4",
+    "connect-history-api-fallback": "^1.6.0",
+    "debug": "^4.3.2",
+    "email-templates": "^8.0.8",
+    "express": "^4.17.1",
+    "express-sslify": "^1.2.0",
+    "feathers-hooks-common": "^5.0.5",
+    "feathers-sequelize": "^6.2.0",
+    "lodash": "^4.17.21",
+    "moment": "^2.29.1",
+    "moment-timezone": "^0.5.33",
+    "mysql2": "^2.3.0",
+    "node-cron": "^3.0.0",
+    "nodemailer": "^6.7.0",
+    "nodemailer-mailgun-transport": "^2.1.3",
+    "pg": "^8.7.1",
+    "pg-hstore": "^2.3.3",
+    "sequelize": "^6.7.0",
+    "typescript": "~4.3.5"
+  },
+  "devDependencies": {
+    "ts-node-dev": "^1.1.8",
+    "rimraf": "^3.0.2"
   }
 }
