+++ packages/ilmomasiina-server/src/services/admin/event/createEvent.ts
@@ -1,53 +1,16 @@
 import _ from 'lodash';
 
+import {
+  AdminEventCreateBody,
+  adminEventCreateEventAttrs,
+  adminEventCreateQuestionAttrs,
+  adminEventCreateQuotaAttrs,
+} from '@tietokilta/ilmomasiina-api/src/services/admin/events/create';
+import { AdminEventGetResponse } from '@tietokilta/ilmomasiina-api/src/services/events/details';
 import { Event } from '../../../models/event';
 import { Question } from '../../../models/question';
 import { Quota } from '../../../models/quota';
-import { AdminEventGetResponse, getEventDetailsForAdmin } from '../../event/getEventDetails';
-
-// Attributes included in POST /api/events for Event instances.
-export const adminEventCreateEventAttrs = [
-  'title',
-  'slug',
-  'date',
-  'registrationStartDate',
-  'registrationEndDate',
-  'openQuotaSize',
-  'description',
-  'price',
-  'location',
-  'webpageUrl',
-  'facebookUrl',
-  'draft',
-  'listed',
-  'signupsPublic',
-  'verificationEmail',
-] as const;
-
-// Attributes included in POST /api/events for Question instances.
-export const adminEventCreateQuestionAttrs = [
-  'question',
-  'type',
-  'options',
-  'required',
-  'public',
-] as const;
-
-// Attributes included in POST /api/events for Quota instances.
-export const adminEventCreateQuotaAttrs = [
-  'title',
-  'size',
-] as const;
-
-// Data type definitions for the request body.
-export interface AdminEventCreateQuestion extends Pick<Question, typeof adminEventCreateQuestionAttrs[number]> {}
-
-export interface AdminEventCreateQuota extends Pick<Quota, typeof adminEventCreateQuotaAttrs[number]> {}
-
-export interface AdminEventCreateBody extends Pick<Event, typeof adminEventCreateEventAttrs[number]> {
-  questions: AdminEventCreateQuestion[];
-  quotas: AdminEventCreateQuota[];
-}
+import { getEventDetailsForAdmin } from '../../event/getEventDetails';
 
 export default async (data: AdminEventCreateBody): Promise<AdminEventGetResponse> => {
   // Pick only allowed attributes and add order
