+++ packages/ilmomasiina-server/src/services/authentication.ts
@@ -5,7 +5,6 @@ import _ from 'lodash';
 
 import config from '../config';
 import { IlmoApplication } from '../defs';
-import { UserDetails } from './user';
 
 // By default, LocalStrategy calls GET /api/users with provider=rest to get the subject.
 // As we disallow that API externally, we need to drop the provider param from the call.
@@ -16,20 +15,6 @@ class CustomLocalStrategy extends LocalStrategy {
     const entity = await super.getEntity(result, _.omit(params, 'provider'));
     return entity && _.pick(entity, 'id', 'email');
   }
-}
-
-// Response schema.
-export interface AuthResponse {
-  accessToken: string;
-  user: UserDetails;
-  authentication: {
-    strategy: string;
-    accessToken: string;
-    payload: {
-      iat: number;
-      exp: number;
-    };
-  };
 }
 
 export default function setupAuthentication(this: IlmoApplication) {
