+++ packages/ilmomasiina-server/src/services/event/getEventDetails.ts
@@ -1,120 +1,22 @@
 import { NotFound } from '@feathersjs/errors';
 import _ from 'lodash';
 
+import {
+  adminEventGetEventAttrs,
+  AdminEventGetResponse,
+  adminEventGetSignupAttrs,
+  eventGetAnswerAttrs,
+  eventGetEventAttrs,
+  eventGetQuestionAttrs,
+  eventGetQuotaAttrs,
+  EventGetResponse,
+  eventGetSignupAttrs,
+} from '@tietokilta/ilmomasiina-api/src/services/events/details';
 import { Answer } from '../../models/answer';
 import { Event } from '../../models/event';
 import { Question } from '../../models/question';
 import { Quota } from '../../models/quota';
 import { Signup } from '../../models/signup';
-
-// Attributes included in GET /api/events/ID for Event instances.
-export const eventGetEventAttrs = [
-  'title',
-  'slug',
-  'date',
-  'registrationStartDate',
-  'registrationEndDate',
-  'openQuotaSize',
-  'description',
-  'price',
-  'location',
-  'webpageUrl',
-  'facebookUrl',
-  'signupsPublic',
-] as const;
-
-// Attributes included in GET /api/admin/events/ID for Event instances.
-export const adminEventGetEventAttrs = [
-  ...eventGetEventAttrs,
-  'id',
-  'draft',
-  'listed',
-  'verificationEmail',
-] as const;
-
-// Attributes included in results for Question instances.
-export const eventGetQuestionAttrs = [
-  'id',
-  'question',
-  'type',
-  'options',
-  'required',
-  'public',
-] as const;
-
-// Attributes included in results for Quota instances.
-export const eventGetQuotaAttrs = [
-  'id',
-  'title',
-  'size',
-] as const;
-
-// Attributes included in GET /api/events/ID for Signup instances.
-export const eventGetSignupAttrs = [
-  'firstName',
-  'lastName',
-  'status',
-  'position',
-  'createdAt',
-] as const;
-
-// Attributes included in GET /api/admin/events/ID for Signup instances.
-export const adminEventGetSignupAttrs = [
-  ...eventGetSignupAttrs,
-  'id',
-  'email',
-] as const;
-
-// Attributes included in results for Answer instances.
-export const eventGetAnswerAttrs = [
-  'questionId',
-  'answer',
-] as const;
-
-// Type definitions for the endpoint: pick the columns above and add relations.
-
-export interface EventGetQuestionItem
-  extends Omit<Pick<Question, typeof eventGetQuestionAttrs[number]>, 'options'> {
-  options: string[] | null;
-}
-
-export interface EventGetAnswerItem extends Pick<Answer, typeof eventGetAnswerAttrs[number]> {}
-
-export interface EventGetSignupItem extends Pick<Signup, typeof eventGetSignupAttrs[number]> {
-  answers: EventGetAnswerItem[];
-}
-
-export interface EventGetQuotaItem extends Pick<Quota, typeof eventGetQuotaAttrs[number]> {
-  signups: EventGetSignupItem[] | null;
-  signupCount: number;
-}
-
-export interface EventGetResponse extends Pick<Event, typeof eventGetEventAttrs[number]> {
-  questions: EventGetQuestionItem[];
-  quotas: EventGetQuotaItem[];
-  millisTillOpening: number | null;
-  registrationClosed: boolean;
-}
-
-// Type definitions for the admin variant of the endpoint.
-
-export type AdminEventGetQuestionItem = EventGetQuestionItem;
-
-export type AdminEventGetAnswerItem = EventGetAnswerItem;
-
-export interface AdminEventGetSignupItem extends Pick<Signup, typeof adminEventGetSignupAttrs[number]> {
-  answers: AdminEventGetAnswerItem[];
-}
-
-export interface AdminEventGetQuotaItem extends Pick<Quota, typeof eventGetQuotaAttrs[number]> {
-  signups: AdminEventGetSignupItem[];
-  signupCount: number;
-}
-
-export interface AdminEventGetResponse extends Pick<Event, typeof adminEventGetEventAttrs[number]> {
-  questions: AdminEventGetQuestionItem[];
-  quotas: AdminEventGetQuotaItem[];
-}
 
 // Admin queries use ids (so that the slug can be safely edited), user queries use slugs.
 type EventGetIdentifier = { id: Event['id'] } | { slug: string };
